Introduce ICompiler
