+++ runtime/onert/api/src/nnfw_api_internal.cc
@@ -17,6 +17,7 @@
 #include "nnfw_api_internal.h"
 #include "CustomKernelRegistry.h"
 #include "compiler/Compiler.h"
+#include "compiler/MultiModelCompiler.h"
 #include "util/ConfigSource.h"
 #include "util/Exceptions.h"
 #include "util/logging.h"
@@ -421,9 +422,18 @@ NNFW_STATUS nnfw_session::prepare()
 
   try
   {
-    auto compiler = std::make_unique<onert::compiler::Compiler>(_nnpkg, _coptions);
-    _nnpkg.reset();
-    _compiler_artifact = compiler->compile();
+    if (_nnpkg->model_count() == 1)
+    {
+      auto compiler = std::make_unique<onert::compiler::Compiler>(_nnpkg, _coptions);
+      _nnpkg.reset();
+      _compiler_artifact = compiler->compile();
+    }
+    else
+    {
+      auto compiler = std::make_unique<onert::compiler::MultiModelCompiler>(_nnpkg, _coptions);
+      _nnpkg.reset();
+      _compiler_artifact = compiler->compile();
+    }
     _execution = std::make_unique<onert::exec::Execution>(_compiler_artifact->_executors);
   }
   catch (const std::exception &e)
